+++ srcs/ft_math/general/ft_floats.c
@@ -1,27 +1,27 @@
 #include <ft_floats.h>
 
-bool	eq_f(float a, float b)
+bool	eq_f(double a, double b)
 {
-	if (fabs(a - b) <= EPSILON)
+	if (fabsl(a - b) <= EPSILON)
 		return (true);
 	return (false);
 }
 
-bool	less_eq_f(float a, float b)
+bool	less_eq_f(double a, double b)
 {
 	if (a < b || eq_f(a, b))
 		return (true);
 	return (false);
 }
 
-bool	greater_eq_f(float a, float b)
+bool	greater_eq_f(double a, double b)
 {
 	if (a > b || eq_f(a, b))
 		return (true);
 	return (false);
 }
 
-float	clamp_f(float nb, float min, float max)
+double	clamp_f(double nb, double min, double max)
 {
-	return (fmax(min, fmin(nb, max)));
+	return (fmaxl(min, fminl(nb, max)));
 }
